Add files via upload
Compatibility Confirmed on Everdrive64-X7 / N64 Console / Analogue 3D / Simple64

Details: 
All 10 Levels Available 
Debug Mode Available (Press Z+R to Enable and L or R to cycle through options and A to Select)

Debug Options: 
Get all stuff = Enable to get all ammo types, full health, and the jetpack (refills each time you use)
(Ammo Types: Seeker Missile, Flame Thrower, Pulse Cannon, Stunner Gun, and Disruptor)

Invincibility = Enable to become invincible. Can be toggled On or Off with A. 

50 Lives = Enable to get 50 lives.

Villains Are/Aren’t Sleepy = This is auto set to Aren't Sleepy, so enemies will attack you. When this is set to Are Sleepy enemies freeze. Enemies can still be damaged when frozen and will die when unfrozen. Bosses such as Boba Fett wont spawn if you have this turned on when you enter the area.
Enable/Disable with pushing Analog Stick Up or Down when Vilnians Are/Aren't Sleepy is highlighted.

Kill Dash = Kill yourself immediately. Does not work if Invincibility is on.

Teleport = Teleports Dash to a random point of current level. Only works with On-foot levels. 

Next Level = Warps the player to the next level.

This Level = Warps the player to the start of the current level.

Prev Level = Warps the player to the previous level. 

Wall Ghost = If set to ON, will disable collision for some walls, doors, ceilings, and objects. Enemies and destructible objects will still be solid, though.